Move ejected expo to "expo-ejected" and restore "expo" to pre-ejected state so both can co-exist. The ejected expo app now exists in a new "expo-ejected", a sibling of "expo". The ejected expo app is needed for app icon, launch image, and native module support. The non-ejected expo app will remain in "expo" in order to support better developer experience for all scenarios that don't require ejection, including better support for sharing/testing apps (using the Expo app), ability to more easily update to latest React Native and Expo, etc.
As much as possible all the code in common between "expo" and "expo-ejected" is reused. Move that common code to "expo/src". Add a "copy-expo-src" script in "expo-ejected" package.json to enable copy reuse of that code by "expo-ejected" as a pre-build step (currently manual). This is not ideal reuse. Should eventually look at (a) creating npm modules for the reused code or (b) rn-cli config to support multiple roots (so we can just include the code from the "expo" directory directly in "expo-ejected" via the root index.js without this extra copy step), or some other approach.

Restore the following "expo" files to the pre-ejected state, so that the non-ejected "expo" can work as it did before: "app.json", "package.json", "yarn.lock". Other minor modifications to "expo" as well to support (a) the root "App.js" importing the real "src/App.js", as "index.js" does in "expo-ejected", (b) moving the App.test.js to a root __tests__ directory, etc.

Note that we should look at options in future to support ejection-in-place, wherein we would automatically reapply our project and other modifications needed to the ejected app after ejecting again. Such ejection-in-place should be idempotent.