Disable auto-updating of JS bundle for Android - Remove generated/AppConstants.java from .gitignore - Update AppConstants.java to set ARE_REMOTE_UPDATES_ENABLED to false (similar to areRemoteUpdatesEnabled in EXShell.plist for iOS, which was already correct). Also make sure the JS bundle for Android is using the correct Expo version (32 instead of 31).
NOTE: AppConstants.java is initially generated when ejecting to ExpoKit, and is then updated during builds, but, not all app.json properties are reflected to the already initially generated AppConstants.java file, which seems like an Expo bug. (We already did have updates disabled in app.json.) After initial ejection it seems only the embedded response for the embedded JS bundle is updated. This happens during 'expo publish'. That published JS bundle is then embedded as a resource in the app. With this change, the Android app (like iOS) should no longer try to auto-update to JS bundles published after the .apk build completed and the JS bundle was embedded in that .apk.

See: https://trello.com/c/sYiATRO0